refactor: refactoring docs
