+++ src/scripts/components/changeCount.js
@@ -7,27 +7,23 @@ const nodes = {
 };
 
 const Counter = function ({ todobox, inprogresBox, doneBox }) {
-  (this.todobox = todobox),
-    (this.inprogresBox = inprogresBox),
-    (this.doneBox = doneBox),
-    (this.changeCount = function () {
-      let count = arrayOfTodos.filter((e) => e.isProgress === "start").length;
-      let countProgress = arrayOfTodos.filter(
-        (e) => e.isProgress === "inProgress"
-      ).length;
+  this.todobox = todobox,
+  this.inprogresBox = inprogresBox,
+  this.doneBox = doneBox,
 
-      let countDone = arrayOfTodos.filter(
-        (e) => e.isProgress === "done"
-      ).length;
-
-      this.todobox.innerText = count;
-      this.inprogresBox.innerText = countProgress;
-      this.doneBox.innerText = countDone;
-    });
+  this.changeCount = function () {
+    let count = arrayOfTodos.filter((e) => e.isProgress === "start").length;
+    let countProgress = arrayOfTodos.filter((e) => e.isProgress === "inProgress").length;
+    let countDone = arrayOfTodos.filter((e) => e.isProgress === "done").length;
+    this.todobox.innerText = count;
+    this.inprogresBox.innerText = countProgress;
+    this.doneBox.innerText = countDone;
+  };
 };
 
 const changeCount = () => {
   let obj = new Counter(nodes);
+  
   return obj.changeCount();
 };
